* develop: (25 commits)
  fixing PR comments
  fixing PR comments
  fixing PR comments
  Update src/Helpers/StoreTrait.php
  adding helper to store
  adding helper
  Update src/Enqueue/Theme/EnqueueThemeCli.php
  Apply suggestions from code review
  Update src/Geolocation/AbstractGeolocation.php
  Apply suggestions from code review
  adding PR fixes
  adding PR fixes
  adding PR fixes
  adding PR fixes
  Update tests/View/EscapedViewCliTest.php
  Apply suggestions from code review
  Apply suggestions from code review
  Apply suggestions from code review
  Apply suggestions from code review
  Apply suggestions from code review
  ...

# Conflicts:
#	src/Cli/Cli.php

+++ tests/Helpers.php
@@ -128,9 +128,12 @@ function setBeforeEach() {
  *
  * @return void
  */
-function setAfterEach() {
+function setAfterEach($delete = true) {
 	Monkey\tearDown();
-	deleteCliOutput();
+
+	if ($delete) {
+		deleteCliOutput();
+	}
 
 	putenv('ES_CLI_SUCCESS_HAPPENED');
 	putenv('ES_CLI_ERROR_HAPPENED');
